backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-decouple==3.8

# Redis & Task Queue
redis==5.0.1
celery==5.3.4

# File Processing & OCR
pillow==10.1.0
pytesseract==0.3.10
python-magic==0.4.27
aiofiles==23.2.1

# HTTP Requests
httpx==0.25.2
aiohttp==3.9.0

# Email & Notifications
fastapi-mail==1.4.1
twilio==8.12.0

# Data Processing & Export
pandas==2.1.4
openpyxl==3.1.2
jinja2==3.1.2

# Validation & Utilities
email-validator==2.1.0
phonenumbers==8.13.25
python-slugify==8.0.1

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Testing dependencies
faker==20.1.0
factory-boy==3.3.0
freezegun==1.2.2